Refactored dpdk component to reflect its role in edgeOS
