+++ src/platform/i386/runscripts/dpdk_init.sh
@@ -1,4 +1,4 @@
 #!/bin/sh
 
-cp cos_dpdk.o llboot.o
+cp network_interface_nf.o llboot.o
 ./cos_linker "llboot.o, :" ./gen_client_stub
